Refactor: initial attempt to split front-end and computing logic
This PR mainly addressed the following problems:

- Current implementation somehow mixed the front-end code and
the actual computing logic. This PR moved all the "back-end"
into `core` folder. However, in future, we should expect those
code to sit in a separated package, like `hedgehog-core`.
- Code cleanup. This PR removed most of the redundant code and
dependencies, and did a whole-project formating. In future we
should use ESlint to force this workflow. An extra PR would be
pleased.

There are some future discussions:

- Move to Yarn. NPM sucks at workspace management, which blocks
the complete separation of `hedgehog-core`. Yarn is also the
recommanded package manager of React.
- It is unusual to add type annotations in JS files nowadays,
and we see both `flow` and `typescript` are presented in deps.
We should either switch to TypeScript internally, or remove the
current type annotations in JS sources.

Signed-off-by: imtsuki <me@qjx.app>